requirements.txt
cupy
numpy
pytest-benchmark

